Lock bootstrap version to avoid incorrect styling for Belop component
Bootstrap was only limited to major version 4, meaning minor version can change.
We use 4.4.1 (locked) in byggeren, but got 4.5.1 in utfylleren because of npm dependency version resolution.
As a result, a CSS change in 4.5.x caused the input for inputs with postfix or suffix to have 1% width.
Hard-coding the bootstrap version for now to avoid this.

Will remove Bootstrap as soon as possible to avoid shit like this from happening, as we really don't need or want bootstrap.